class 3

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (req) => {
+    await req.jwtVerify();
+  });
+
+  app.get("/memories", async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -19,18 +26,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (req) => {
+  app.get("/memories/:id", async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(req.params);
 
-    const memory = prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return res.status(401).send();
+    }
 
     return memory;
   });
@@ -49,19 +60,29 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "4b0a9bcc-cb55-490f-afd8-6fa87a197909",
+        userId: req.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.delete("/memories/:id", async (req) => {
+  app.delete("/memories/:id", async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(req.body);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
@@ -70,7 +91,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.put("/memories/:id", async (req) => {
+  app.put("/memories/:id", async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -85,7 +106,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
